WIP refactoring as a new rosidl generator
