+++ rosidl_generator_type_hash/bin/rosidl_generator_type_hash
@@ -5,20 +5,20 @@ import os
 import sys
 
 try:
-    from rosidl_generator_c import generate_c
+    from rosidl_generator_type_hash import generate_type_hash
 except ImportError:
     # modifying sys.path and importing the Python package with the same
     # name as this script does not work on Windows
-    rosidl_generator_c_root = os.path.dirname(os.path.dirname(__file__))
-    rosidl_generator_c_module = os.path.join(
-        rosidl_generator_c_root, 'rosidl_generator_c', '__init__.py')
-    if not os.path.exists(rosidl_generator_c_module):
+    rosidl_generator_type_hash_root = os.path.dirname(os.path.dirname(__file__))
+    rosidl_generator_type_hash_module = os.path.join(
+        rosidl_generator_type_hash_root, 'rosidl_generator_type_hash', '__init__.py')
+    if not os.path.exists(rosidl_generator_type_hash_module):
         raise
     from importlib.machinery import SourceFileLoader
 
-    loader = SourceFileLoader('rosidl_generator_c', rosidl_generator_c_module)
-    rosidl_generator_c = loader.load_module()
-    generate_c = rosidl_generator_c.generate_c
+    loader = SourceFileLoader('rosidl_generator_type_hash', rosidl_generator_type_hash_module)
+    rosidl_generator_type_hash = loader.load_module()
+    generate_type_hash = rosidl_generator_type_hash.generate_type_hash
 
 
 def main(argv=sys.argv[1:]):
@@ -31,7 +31,7 @@ def main(argv=sys.argv[1:]):
         help='The location of the file containing the generator arguments')
     args = parser.parse_args(argv)
 
-    generate_c(args.generator_arguments_file)
+    generate_type_hash(args.generator_arguments_file)
 
 
 if __name__ == '__main__':
